requirements/dev.txt
# Include base requirements
-r base.txt

# Testing
pytest==7.4.3
pytest-django==4.9.0
pytest-cov==6.1.0
pytest-xdist==3.6.3
factory-boy==3.3.0
coverage==7.11.0

# Code Quality
black==25.9.0
ruff==0.9.7
mypy==1.18.2
flake8==7.2.2
autopep8==2.3.2
isort==6.1.7

# Type Stubs
django-stubs==5.2.7
djangorestframework-stubs==3.16.5
pandas-stubs==2.3.2.250926
types-PyYAML==6.0.12.20250312
types-requests==2.32.0.20250821
types-redis==4.6.0.20250818

# Security Scanning
bandit==1.8.6

# Development Tools
ipython==9.2.0
ipdb==0.14.2
django-extensions==3.3.1
django-debug-toolbar==5.1.2
